refactor(test): samle tester i test/unit og test/integration
Flytter enhetstester fra src/ til test/unit/ med en struktur som
speiler src/ (JVM/Gradle-stil). Playwright-specs flyttes fra e2e/ til
test/integration/.

- test/unit/**: Vitest-enhetstester importerer via @src/*-alias
- test/integration/**: Playwright-specs med testMatch **/*.spec.ts
- vitest.config.ts: include → test/unit/**/*.test.{ts,tsx}
- playwright.config.ts: testDir → ./test/integration
