+++ test/integration/i18n-redirect.spec.ts
